pyproject.toml
[tool.poetry]
name = "electricity-consumption"
version = "0.1.0"
description = ""
authors = ["Sofia Valenti"]
readme = "README.md"
packages = [{include = "electricity_consumption"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
responses = "^0.22.0"
mockito = "^1.4.0"
requests = "^2.28.2"
pytest = "^7.2.1"
pandas = "^1.5.3"
matplotlib = "^3.6.3"
python-dotenv = "^0.21.1"
coverage = "^7.2.1"
logging = "^0.4.9.6"
edistribucionapi = {git = "https://github.com/trocotronic/edistribucion"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src/electricity_consumption"
]